requirements.txt
streamlit
pandas
xlsxwriter
supabase
python-dotenv
google-generativeai
